refactoring for lib modules
