+++ src/lib/dom/html/index.ts
@@ -1,0 +1,1 @@
+export const getHTML = () => document.documentElement.innerHTML;
