+++ src/lib/dom/nodes/index.ts
@@ -1,0 +1,2 @@
+export const getTextOfChildNode = (parentNode: Cheerio, childNode: string) =>
+  parentNode.find(childNode).text();
