+++ src/lib/dom/window_scroll/index.ts
@@ -1,8 +1,6 @@
 import { Page } from 'puppeteer';
 
-export const sleep = (ms: number) => {
-  return new Promise(resolve => setTimeout(resolve, ms));
-};
+import { sleep } from '../../sleep';
 
 export const autoScroll = async (page: Page) => {
   await page.evaluate(async () => {
@@ -39,28 +37,6 @@ export const scrollWithCount = async (page: Page, scrollCount: number) => {
   }
 };
 
-export const getTextOfChildNode = (parentNode: Cheerio, childNode: string) =>
-  parentNode.find(childNode).text();
-
-export const getNormalizedThousandthValue = (value: string) => {
-  let normalizedValue = Number(value);
-
-  if (value.includes('K')) {
-    const valueWithoutKFormat = Number(value.replace('K', ''));
-    normalizedValue = valueWithoutKFormat * 1000;
-  }
-
-  return normalizedValue;
-};
-
-export const getHTML = () => document.documentElement.innerHTML;
-
 export const scrollWithDefaultYDiapason = () => {
   window.scrollBy(0, 800);
-};
-
-export const getTextWithAlphaOnly = (text: string) => {
-  const textWithAlphaOnly = text.replace(/[^a-zA-Z\s]+/g, '');
-
-  return textWithAlphaOnly;
 };
