+++ src/lib/sleep/index.ts
@@ -1,0 +1,3 @@
+export const sleep = (ms: number) => {
+  return new Promise(resolve => setTimeout(resolve, ms));
+};
